Изменил структуру проекта Создал директорию Common. В ней создал файл required_date и перенёс туда одноименнный класс
